+++ Examples/beacons/Examples/ESTAppDelegate.h
@@ -13,6 +13,8 @@
 
 @property (strong, nonatomic) UIWindow *window;
 @property (strong, nonatomic) UINavigationController *mainNavigation;
+@property (strong, nonatomic) ESTCloudManager *cloudManager;
+
 
 @end